+++ src/main/java/com/saludSystem/repositories/modules/Configuration/SedeRepository.java
@@ -3,5 +3,5 @@ package com.saludSystem.repositories.modules.Configuration;
 import com.saludSystem.entities.configuracion.Sede;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SedeRepository extends JpaRepository<Sede, Long> {
+public interface SedeRepository extends JpaRepository<Sede, Integer> {
 }
