+++ src/main/java/com/saludSystem/repositories/modules/Configuration/TipoDocumentoRepository.java
@@ -3,5 +3,5 @@ package com.saludSystem.repositories.modules.Configuration;
 import com.saludSystem.entities.configuracion.TipoDocumento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TipoDocumentoRepository extends JpaRepository<TipoDocumento, Long> {
+public interface TipoDocumentoRepository extends JpaRepository<TipoDocumento, Integer> {
 }
